requirements.txt
Django==3.2.10
django-cors-headers==3.11.0
djangorestframework==3.13.1
openpyxl==2.4.11
pandas==1.4.3
